- Exhumed: Perform weapon bobbing with maximum precision and implement `cl_weaponsway`.

+++ source/exhumed/src/aistuff.h
@@ -152,7 +152,7 @@ void StopFiringWeapon(short nPlayer);
 void FireWeapon(short nPlayer);
 void CheckClip(short nPlayer);
 void MoveWeapons(short nPlayer);
-void DrawWeapons(int smooth);
+void DrawWeapons(double smooth);
 
 // items
